Put email settings in a separate component.
